+++ src/Products/urban/migration/update_260.py
@@ -1,4 +1,8 @@
 from plone import api
+from Products.urban.profiles.extra.config_default_values import default_values
+from Products.urban.setuphandlers import createVocabularies
+from Products.urban.setuphandlers import createVocabularyFolders
+
 import logging
 
 
@@ -12,3 +16,15 @@ def add_couple_to_preliminary_notice(context):
     setup_tool.runImportStepFromProfile('profile-Products.urban:preinstall', 'typeinfo')
     setup_tool.runImportStepFromProfile('profile-Products.urban:preinstall', 'workflow')
     logger.info("upgrade step done!")
+
+
+def add_vocabularies_to_preliminary_notice_foldercategories(context):
+    portal_urban = api.portal.get_tool('portal_urban')
+    config_folder = getattr(portal_urban, 'preliminarynotice')
+    preliminary_notice_vocabularies = default_values['PreliminaryNotice']
+    createVocabularyFolders(
+        container=config_folder, vocabularies=preliminary_notice_vocabularies, site=None
+    )
+    createVocabularies(
+        container=config_folder, vocabularies=preliminary_notice_vocabularies
+    )